requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
chardet==3.0.4
greenlet==1.0.0
idna==3.1
Mako==1.1.4
MarkupSafe==1.1.1
multidict==5.1.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.4.0b3
typing-extensions==3.7.4.3
yarl==1.6.3
aiohttp==3.7.3
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
chardet==3.0.4
greenlet==1.0.0
idna==3.1
Mako==1.1.4
MarkupSafe==1.1.1
multidict==5.1.0
PyJWT==2.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.4.0b3
typing-extensions==3.7.4.3
yarl==1.6.3
